Android: fix narrowing int->uint in NetherReactorPattern
c++_static enforces C++11 narrowing; change local tile ID vars to
unsigned int so they match the unsigned int[3][3][3] array type.

+++ handheld/src/world/level/tile/NetherReactorPattern.cpp
@@ -1,9 +1,9 @@
 #include "NetherReactorPattern.h"
 #include "Tile.h"
 NetherReactorPattern::NetherReactorPattern( ) {
-	const int goldId = Tile::goldBlock->id;
-	const int stoneId = Tile::stoneBrick->id;
-	const int netherCoreId = Tile::netherReactor->id;
+	const unsigned int goldId = (unsigned int)Tile::goldBlock->id;
+	const unsigned int stoneId = (unsigned int)Tile::stoneBrick->id;
+	const unsigned int netherCoreId = (unsigned int)Tile::netherReactor->id;
 	const unsigned int types[3][3][3] =
 	{
 		// Level 0
